2 Console Tools
I wrote 2 small tools. And I’d like to share them with you.

One is called rc which could be used for recording the output of a
command into a file with printing them to console at the same time.
C:\scripts>rc
Usage:
rc <command>

Another tool is called loop. It could be used for executing some command
repeatedly. You could specify the loop times and the duration.
C:\scripts>loop
Usage:
loop -cmd "[command]" <-times [number of loop times]> <-sleep [seconds]>
<-log>
loop -c "[command]" <-t [number of loop times]> <-s [seconds]> <-l>
loop [command]

The valid number of loop times is between 1 and 1000
The valid value of sleep is between 1 second and 180 seconds
Switch "-l" or "-log" is used for enabling the log

Just for fun.

Please have a try.

Both binary files and source code are attached to the mail. Rename the
tools_zip to tools.zip.

Comments are welcome.

And currently it is not supported to execute them together.